Re-classification of the package creating folders for passive-component and structure constants and special function are put in the special.py file
